[ASTGen] Move source manager diagnostics code into its own file
